Fixed incorrect steps in existing H2_UMV example file
